Refactor: Use clean package-relative imports for morty module
- Replace dynamic file loader and try/except fallback imports with idiomatic package-relative imports
- Import from .morty_debug_plots and .morty_helper using relative imports
- Require invocation as 'python -m morty.morty' from project root
- Update README with single, clean invocation method
- Create morty/__init__.py to mark morty as a package
